splits the script into modules
