--- js/functions.js
@@ -1,37 +1,0 @@
-const checkStringLength = (string, length) => string.length <= length;
-
-checkStringLength('проверяемая строка', 20); // true
-checkStringLength('проверяемая строка', 18); // true
-checkStringLength('проверяемая строка', 10); // false
-
-const isPalindrome = (string) => {
-  string = string.replaceAll(' ', '').toLowerCase();
-
-  for (let i = 0; i < Math.floor(string.length / 2); i++) {
-    if (string[i] !== string[string.length - 1 - i]) {
-      return false;
-    }
-  }
-
-  return true;
-};
-
-isPalindrome('топот'); // true
-isPalindrome('ДовОд'); // true
-isPalindrome('Кекс'); // false
-isPalindrome('Лёша на полке клопа нашёл '); // true
-
-const extractNumber = (string) => {
-  string = String(string).replace(/\D/g, '');
-
-  return parseInt(string, 10);
-};
-
-extractNumber('2023 год'); // 2023
-extractNumber('ECMAScript 2022'); // 2022
-extractNumber('1 кефир, 0.5 батона'); // 105
-extractNumber('агент 007'); // 7
-extractNumber('а я томат'); // NaN
-extractNumber(2023); // 2023
-extractNumber(-1); // 1
-extractNumber(1.5); // 15
